Can create post + see post
adding photos to post still not working as intended

+++ yumelinkapp/views/create_edit_post_view.py
@@ -1,8 +1,9 @@
+from django.contrib import messages
 from django.http import HttpResponseRedirect
 from django.shortcuts import get_object_or_404
-from django.urls import reverse_lazy
+from django.urls import reverse_lazy, reverse
 from django.views.generic.edit import UpdateView
-from yumelinkapp.models import Post, PostImage
+from yumelinkapp.models import Post, PostImage, User
 from yumelinkapp.forms import PostImageFormSet
 
 
@@ -13,7 +14,6 @@ class CreateEditPostView(UpdateView):
     model = Post
     template_name = 'yumelink/create_edit_post.html'
     fields = ['content', 'filter_content']
-    success_url = reverse_lazy('yumelinkapp:home')
 
     def get_object(self, queryset=None):
         post_id = self.kwargs.get('pk')
@@ -22,33 +22,52 @@ class CreateEditPostView(UpdateView):
         return None
 
     def form_valid(self, form):
-        if not self.get_object():
-            form.instance.author = self.request.user
-        post = form.save()
+        form.instance.user = User.objects.get(id=self.request.user.id)
+
+        self.object = form.save()
 
         context = self.get_context_data()
         image_formset = context['image_formset']
+
+        if not image_formset:
+            return HttpResponseRedirect(reverse('yumelinkapp:post', kwargs={"pk": self.object.id}))
+
         if image_formset.is_valid():
-            post = form.save()
             images = image_formset.save(commit=False)
             for image in images:
-                image.post = post
-                image.save()
+                image.post = self.object
+                if image.delete_image:
+                    image.delete()
+                else:
+                    image.save()
             for obj in image_formset.deleted_objects:
                 obj.delete()
 
-            return HttpResponseRedirect(self.success_url)
+            return HttpResponseRedirect(reverse('yumelinkapp:post', kwargs={"pk": self.object.id}))
         else:
-            return self.render_to_response(self.get_context_data(form=form))
+            if image_formset.errors or image_formset.management_form.errors:
+                messages.info(self.request, f"Image formset errors:{image_formset.errors}")
+                messages.info(self.request, f"Management form errors:{image_formset.management_form.errors}")
+
+                messages.error(self.request, "failed to post, please try again.")
+
+                return HttpResponseRedirect(reverse('yumelinkapp:edit_post', kwargs={"pk": self.object.id}))
+        return HttpResponseRedirect(reverse('yumelinkapp:post', kwargs={"pk": self.object.id}))
 
     def get_context_data(self, **kwargs):
         context = super().get_context_data(**kwargs)
         context['user'] = self.request.user
         context['post'] = self.object
-        if self.request.method == 'POST':
-            context['image_formset'] = PostImageFormSet(self.request.POST, self.request.FILES,
-                                                        queryset=PostImage.objects.filter(post=self.object))
-        else:
-            context['image_formset'] = PostImageFormSet(queryset=PostImage.objects.filter(post=self.object))
+        context['image_formset'] = PostImageFormSet(queryset=PostImage.objects.filter(post=self.object))
 
         return context
+
+    def get_form_kwargs(self):
+        kwargs = super().get_form_kwargs()
+        if self.request.method == 'POST':
+            kwargs['data'] = self.request.POST
+            kwargs['files'] = self.request.FILES
+        else:
+            kwargs['data'] = None
+            kwargs['files'] = None
+        return kwargs
